+++ src/hydrate-with-store.jsx
@@ -1,8 +1,9 @@
 import { fromJS } from 'immutable';
 import { hydrate } from 'react-dom';
 import { Provider } from 'react-redux';
+import { BrowserRouter } from 'react-router-dom';
 
-import { PLACEHOLDER, PRELOADED_STATE } from './constants';
+import { PLACEHOLDER, PRELOADED_STATE, PRELOADED_STATE_PLACEHOLDER_ID } from './constants';
 import createStore from './create-store';
 
 export default (Component, reducers, middlewares, inDevelopment, projectInitType) => {
@@ -10,8 +11,21 @@ export default (Component, reducers, middlewares, inDevelopment, projectInitType
     if (placeholder) {
         const initialState = window[PRELOADED_STATE];
         const store = createStore(reducers, fromJS(initialState), middlewares, inDevelopment, projectInitType);
-        hydrate(<Provider store={store}><Component /></Provider>, placeholder);
+        const jsx = (
+            <Provider store={store}>
+                <BrowserRouter>
+                    <Component />
+                </BrowserRouter>
+            </Provider>
+        );
+
+        hydrate(jsx, placeholder);
         delete window[PRELOADED_STATE];
+
+        const scriptTag = document.querySelector(`#${PRELOADED_STATE_PLACEHOLDER_ID}`);
+        if (scriptTag) {
+            scriptTag.remove();
+        }
     } else {
         console.error(`Cannot find placeholder '#${PLACEHOLDER}'.`);
     }